FoodZilla/requirements.txt
numpy
cython
pystan
fbprophet
django
pillow
opencv-python
matplotlib
seaborn
